+++ src/Data_presentation.h
@@ -13,7 +13,7 @@ class Data_presentation
   void display_save_data();
   void presentation_measurements_ph(const float temperature, const float ph);
   void presentation_measurements_ec(const float temperature, const float ec);
-  void display_calibration_ph(const int sample, const float temperature);
+  void display_calibration_ph(const uint8_t sample, const float temperature);
   void display_calibration_ec(const double sample, const int position, const float temperature);
 
   private:
